backend_python/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-dotenv
passlib[bcrypt]
emails
